Added the general backbone of the program
Fixed the fact that the previous stuff wouldn't even compile because I forgot to put in the RDA define.

Added the general state system. Needs: Real time clock outputs, water monitoring, temperature monitoring. ISR for disabled
Currently has: step motor at button press in correct states (Needs to output changes though). State changes at insufficient water level, temperature, disable button press.

Added 'fanset' function, which accepts a bool. With 'true', the fan is turned on and  'on' is output to the serial. with 'false', the fan is turned off and  'off' is output  to the serial. Needs Real time clock added.

Added 'led_set' function. Reading that, I realize it is not consistent with my other function. Whoops. Anyway, it is supposed to take an int, and depending on if the int is  0, 1, 2, or 3, it turns on one of the LEDs and the other ones off. Not implemented yet. Was planning on using switch-case, perhaps with a default if something wrong is put in... though that shouldn't happen.